Added the main pipeline of the code. Created utility files with different functionalities and debugging purpuses.
