+++ ChessDetector/README.md
